+++ packages/core/admin/admin/src/components/SearchInput.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 
 import { Icon, IconButton, Searchbar, SearchForm } from '@strapi/design-system';
-import { useQueryParams } from '@strapi/helper-plugin';
 import { Search as SearchIcon } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 
 import { TrackingEvent, useTracking } from '../features/Tracking';
+import { useQueryParams } from '../hooks/useQueryParams';
 
 interface SearchInputProps {
   disabled?: boolean;
