+++ packages/core/admin/admin/src/content-manager/components/LeftMenu.tsx
@@ -8,12 +8,12 @@ import {
   SubNavSection,
   SubNavSections,
 } from '@strapi/design-system/v2';
-import { useQueryParams } from '@strapi/helper-plugin';
 import { parse, stringify } from 'qs';
 import { useIntl } from 'react-intl';
 import { NavLink } from 'react-router-dom';
 
 import { useTypedSelector } from '../../core/store/hooks';
+import { useQueryParams } from '../../hooks/useQueryParams';
 import { getTranslation } from '../utils/translations';
 
 const LeftMenu = () => {
