+++ packages/core/admin/admin/src/content-manager/pages/EditView/components/FormInputs/UID.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import { FieldAction, Flex, TextInput, Typography, useComposedRefs } from '@strapi/design-system';
-import { useQueryParams } from '@strapi/helper-plugin';
 import { CheckCircle, ExclamationMarkCircle, Loader, Refresh } from '@strapi/icons';
 import { Contracts } from '@strapi/plugin-content-manager/_internal/shared';
 import { useIntl } from 'react-intl';
@@ -12,6 +11,7 @@ import { useNotification } from '../../../../../features/Notifications';
 import { useAPIErrorHandler } from '../../../../../hooks/useAPIErrorHandler';
 import { useDebounce } from '../../../../../hooks/useDebounce';
 import { useFocusInputField } from '../../../../../hooks/useFocusInputField';
+import { useQueryParams } from '../../../../../hooks/useQueryParams';
 import { useDoc } from '../../../../hooks/useDocument';
 import {
   useGenerateUIDMutation,
