+++ packages/core/admin/admin/src/content-manager/pages/ListView/components/TableActions.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 
 import { Button, Flex } from '@strapi/design-system';
 import { LinkButton } from '@strapi/design-system/v2';
-import { useQueryParams } from '@strapi/helper-plugin';
 import { Duplicate, Pencil } from '@strapi/icons';
 import { Contracts } from '@strapi/plugin-content-manager/_internal/shared';
 import { stringify } from 'qs';
@@ -13,6 +12,7 @@ import styled from 'styled-components';
 import { DescriptionComponentRenderer } from '../../../../components/DescriptionComponentRenderer';
 import { useNotification } from '../../../../features/Notifications';
 import { useStrapiApp } from '../../../../features/StrapiApp';
+import { useQueryParams } from '../../../../hooks/useQueryParams';
 import { isBaseQueryError } from '../../../../utils/baseQuery';
 import { useDocumentRBAC } from '../../../features/DocumentRBAC';
 import { Document, useDoc } from '../../../hooks/useDocument';
