+++ packages/core/admin/admin/src/content-manager/pages/ListView/components/ViewSettingsMenu.tsx
@@ -10,7 +10,7 @@ import {
   useCollator,
 } from '@strapi/design-system';
 import { LinkButton } from '@strapi/design-system/v2';
-import { useQueryParams, useRBAC } from '@strapi/helper-plugin';
+import { useRBAC } from '@strapi/helper-plugin';
 import { Cog, Layer } from '@strapi/icons';
 import { stringify } from 'qs';
 import { useIntl } from 'react-intl';
@@ -19,6 +19,7 @@ import styled from 'styled-components';
 
 import { useTypedSelector } from '../../../../core/store/hooks';
 import { useTracking } from '../../../../features/Tracking';
+import { useQueryParams } from '../../../../hooks/useQueryParams';
 import { useDoc } from '../../../hooks/useDocument';
 import { useDocumentLayout } from '../../../hooks/useDocumentLayout';
 import { checkIfAttributeIsDisplayable } from '../../../utils/attributes';
