+++ packages/core/admin/admin/src/index.ts
@@ -16,19 +16,12 @@ export * from './components/Pagination';
 export * from './components/SearchInput';
 export * from './components/Table';
 
+export { useGuidedTour } from './components/GuidedTour/Provider';
+
 /**
  * Features
  */
 export { useTracking, type TrackingEvent } from './features/Tracking';
-
-/**
- * Hooks
- */
-export { useInjectReducer } from './hooks/useInjectReducer';
-// TODO: Replace this export with the same hook exported from the @strapi/admin/strapi-admin/ee in another iteration of this solution
-export { useLicenseLimits } from '../../ee/admin/src/hooks/useLicenseLimits';
-export { useGuidedTour } from './components/GuidedTour/Provider';
-export { useAPIErrorHandler } from './hooks/useAPIErrorHandler';
 export { useStrapiApp, type StrapiAppContextValue } from './features/StrapiApp';
 export {
   useNotification,
@@ -37,6 +30,15 @@ export {
   NotificationsProvider,
 } from './features/Notifications';
 export { useAppInfo, type AppInfoContextValue } from './features/AppInfo';
+
+/**
+ * Hooks
+ */
+export { useInjectReducer } from './hooks/useInjectReducer';
+// TODO: Replace this export with the same hook exported from the @strapi/admin/strapi-admin/ee in another iteration of this solution
+export { useLicenseLimits } from '../../ee/admin/src/hooks/useLicenseLimits';
+export { useAPIErrorHandler, type ApiError } from './hooks/useAPIErrorHandler';
+export { useQueryParams } from './hooks/useQueryParams';
 
 /**
  * Types
@@ -57,7 +59,6 @@ export type {
   HeaderActionDescription,
   HeaderActionProps,
 } from './core/apis/content-manager';
-export type { ApiError } from './types/errors';
 
 /**
  * Utils
