+++ packages/core/admin/admin/src/pages/Marketplace/MarketplacePage.tsx
@@ -13,7 +13,6 @@ import {
   TabPanels,
   Tabs,
 } from '@strapi/design-system';
-import { useQueryParams } from '@strapi/helper-plugin';
 import { ExternalLink, GlassesSquare } from '@strapi/icons';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
@@ -26,6 +25,7 @@ import { useAppInfo } from '../../features/AppInfo';
 import { useNotification } from '../../features/Notifications';
 import { useTracking } from '../../features/Tracking';
 import { useDebounce } from '../../hooks/useDebounce';
+import { useQueryParams } from '../../hooks/useQueryParams';
 
 import { NpmPackagesFilters } from './components/NpmPackagesFilters';
 import { NpmPackagesGrid } from './components/NpmPackagesGrid';
