--- packages/core/admin/admin/src/types/errors.ts
@@ -1,14 +1,0 @@
-import type { errors } from '@strapi/utils';
-
-export type ApiError =
-  | errors.ApplicationError
-  | errors.ForbiddenError
-  | errors.NotFoundError
-  | errors.NotImplementedError
-  | errors.PaginationError
-  | errors.PayloadTooLargeError
-  | errors.PolicyError
-  | errors.RateLimitError
-  | errors.UnauthorizedError
-  | errors.ValidationError
-  | errors.YupValidationError;
