+++ packages/core/admin/admin/src/utils/baseQuery.ts
@@ -3,7 +3,7 @@ import { BaseQueryFn } from '@reduxjs/toolkit/query';
 import { getFetchClient } from '@strapi/helper-plugin';
 import { isAxiosError, type AxiosRequestConfig } from 'axios';
 
-import type { ApiError } from '../types/errors';
+import type { ApiError } from '../hooks/useAPIErrorHandler';
 
 /* -------------------------------------------------------------------------------------------------
  * Axios data
