+++ packages/core/content-releases/admin/src/pages/ReleasesPage.tsx
@@ -8,6 +8,7 @@ import {
   useTracking,
   useAPIErrorHandler,
   useNotification,
+  useQueryParams,
 } from '@strapi/admin/strapi-admin';
 import {
   Alert,
@@ -30,7 +31,7 @@ import {
   Typography,
 } from '@strapi/design-system';
 import { Link } from '@strapi/design-system/v2';
-import { useQueryParams, useRBAC } from '@strapi/helper-plugin';
+import { useRBAC } from '@strapi/helper-plugin';
 import { EmptyDocuments, Plus } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 import { useNavigate, useLocation } from 'react-router-dom';
