+++ packages/core/helper-plugin/MIGRATION_GUIDE.md
@@ -606,6 +606,18 @@ This hook has been removed and not replaced. If you feel like you need this hook
 
 This hook has been removed and not replaced. If you feel like you need this hook, please open an issue on the Strapi repository to discuss your usecase.
 
+### useQueryParams
+
+This hook has been moved. You should import it from the `@strapi/strapi/admin` package:
+
+```tsx
+// Before
+import { useQueryParams } from '@strapi/helper-plugin';
+
+// After
+import { useQueryParams } from '@strapi/strapi/admin';
+```
+
 ## Icons
 
 ### SortIcon
