+++ packages/core/helper-plugin/src/index.ts
@@ -10,7 +10,6 @@ export * from './features/RBAC';
 
 export * from './hooks/useFetchClient';
 export * from './hooks/usePersistentState';
-export * from './hooks/useQueryParams';
 export * from './hooks/useRBAC';
 export * from './hooks/useSelectionState';
