+++ packages/plugins/i18n/admin/src/components/LocalePicker.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { SingleSelect, SingleSelectOption } from '@strapi/design-system';
-import { useQueryParams } from '@strapi/helper-plugin';
+import { useQueryParams } from '@strapi/strapi/admin';
 import { useIntl } from 'react-intl';
 
 import { useI18n } from '../hooks/useI18n';
