+++ packages/plugins/users-permissions/admin/src/pages/Roles/pages/ListPage/index.jsx
@@ -17,7 +17,7 @@ import {
   useFilter,
 } from '@strapi/design-system';
 import { LinkButton } from '@strapi/design-system/v2';
-import { useQueryParams, useRBAC } from '@strapi/helper-plugin';
+import { useRBAC } from '@strapi/helper-plugin';
 import { Plus } from '@strapi/icons';
 import {
   ConfirmDialog,
@@ -26,6 +26,7 @@ import {
   SearchInput,
   BackButton,
   useNotification,
+  useQueryParams,
 } from '@strapi/strapi/admin';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
